requirements.txt
streamlit>=1.31.0,<2
streamlit-pdf==2.0.1
langchain>=0.3.0,<0.4
langchain-core>=0.3.0,<0.4
langchain-openai>=0.2.0,<0.4
langchain-community>=0.3.0,<0.4
langchain-text-splitters>=0.3.0,<0.4
chromadb>=0.5.0,<1
pydantic>=2.5.0,<3
python-dotenv>=1.0.0
pypdf>=4.0.0
python-docx>=1.1.0
Pillow>=10.0.0
pandas>=2.0.0
pytesseract>=0.3.10
